Refactor: 기존 콘솔 뷰 로직을 view/console로 이동
