+++ src/view/console/input-controller.js
