+++ src/view/console/input.js
