+++ src/view/console/output-controller.js
