+++ src/view/console/output.js
